1 задание: вынесла Enum из класса Enimal, переделала на substring ген код потомтсва, сделала в размножении проверку живое ли живтоное. Вопрос: теперь в методе reproduction (класс Animal) в if (isAlive1 == true && asAlive2 == true) - плохая практика? как избежать?

+++ src/task2_1/Cow.java
@@ -3,7 +3,7 @@ package task2_1;
 import java.util.UUID;
 
 public class Cow extends Herbivore{
-    protected Cow(int weight, int age, Gender gender, UUID geneticCode) {
+    protected Cow(float weight, int age, Gender gender, UUID geneticCode) {
         super(weight, age, gender, geneticCode);
         System.out.println("Корова");
     }
