+++ src/task2_1/Lion.java
@@ -3,7 +3,7 @@ package task2_1;
 import java.util.UUID;
 
 public class Lion extends Predator{
-    protected Lion(int weight, int age, Gender gender, UUID geneticCode) {
+    protected Lion(float weight, int age, Gender gender, UUID geneticCode) {
         super(weight, age, gender, geneticCode);
         System.out.println("Лев");
     }
