+++ src/task2_1/Sheep.java
@@ -3,7 +3,7 @@ package task2_1;
 import java.util.UUID;
 
 public class Sheep extends Herbivore{
-    protected Sheep(int weight, int age, Gender gender, UUID geneticCode) {
+    protected Sheep(float weight, int age, Gender gender, UUID geneticCode) {
         super(weight, age, gender, geneticCode);
         System.out.println("Îâöà");
     }
