+++ src/task2_1/Wolf.java
@@ -3,7 +3,7 @@ package task2_1;
 import java.util.UUID;
 
 public class Wolf extends Predator {
-    protected Wolf(int weight, int age, Gender gender, UUID geneticCode) {
+    protected Wolf(float weight, int age, Gender gender, UUID geneticCode) {
         super(weight, age, gender, geneticCode);
         System.out.println("Волк");
     }
